# openapi.yaml
openapi: 3.0.3
info:
  title: tr-engine API
  version: 0.9.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Backend service for trunk-recorder data ingestion and querying.
    Provides REST APIs for accessing radio system data, calls, talkgroups,
    units, and recorder state.

    ## Authentication

    All endpoints (except `/health` and `/auth-init`) require a bearer token
    via the `Authorization: Bearer <token>` header or `?token=` query param.

    **Two-tier auth (optional):** When `WRITE_TOKEN` is configured, mutating
    operations (POST, PATCH, PUT, DELETE) require the write token. Read
    operations (GET) accept either the read token (`AUTH_TOKEN`) or write
    token. If `WRITE_TOKEN` is not set, all operations accept `AUTH_TOKEN`
    (backwards compatible).

    The `/auth-init` endpoint returns only the read token — the write token
    is never exposed by any endpoint.

    ## Data Model: Systems and Sites

    The data model separates **logical radio networks** from **recording
    sites**:

    - **System** (`system_id`): a logical radio network. For P25, this
      is the WACN-level network identified by `(sysid, wacn)`. For
      conventional, it is identified by `(instance_id, sys_name)`.
      Talkgroups and units belong to the system level because they are
      shared across all sites on the same P25 network.

    - **Site** (`site_id`): a recording point within a system. Each
...
      specific site.

...
    ```
...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /call-groups:
    get:
      operationId: listCallGroups
      summary: List call groups
      description: |
        Returns deduplicated call groups (calls from multiple recorders
        grouped together). Each group represents one logical transmission
        captured by one or more recording sites.
      tags: [call-groups]
      parameters:
        - name: sysid
          in: query
          description: Filter by P25 SYSID (comma-separated for multiple)
          schema:
            type: string
        - name: tgid
          in: query
          description: Filter by talkgroup ID (comma-separated for multiple)
          schema:
            type: string
            example: "9178,5344"
        - $ref: "#/components/parameters/startTime"
        - $ref: "#/components/parameters/endTime"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallGroupListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /call-groups/{id}:
    get:
      operationId: getCallGroup
      summary: Get a call group
      description: Returns a call group with all its individual call recordings.
      tags: [call-groups]
      parameters:
        - name: id
          in: path
          required: true
          description: Call group ID
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallGroupDetailResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ----------------------------------------------------------
  # Stats
  # ----------------------------------------------------------
...
  /admin/systems/merge:
    post:
      operationId: mergeSystems
      summary: Merge two systems
      description: |
        Merges all data from `source_id` into `target_id`, then deletes
        the source system. Use this to fix fragmentation caused by
        instance_id or sys_name changes, or to consolidate an unknown
        system after identification.

        **What happens:**
        - All calls under source are repointed to target system_id
          (call_ids are database PKs and do not change)
        - Talkgroups: source talkgroups are merged into target. If both
          have the same tgid, target metadata is kept and call counts are
          combined. Source-only talkgroups are moved to target.
        - Units: same merge logic as talkgroups
        - Unit events: repointed to target system_id
        - Call groups may be consolidated (recordings that now share
          system_id + tgid + start_time get grouped)
        - Source system is deleted

        **Implications:**
        - Call IDs (database PKs) are stable — bookmarks don't break
        - Talkgroup and unit composite IDs change (old system_id prefix
          becomes invalid, new system_id prefix is correct)
        - This operation is **not reversible**

        After merging, update the trunk-recorder config so future MQTT
        messages use the target system's `(instance_id, short_name)`.
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemMergeRequest"
      responses:
        "200":
          description: Merge completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemMergeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Source or target system not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Merge would cause data conflicts that cannot be auto-resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"

# ============================================================
# COMPONENTS
# ============================================================
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token authentication. Pass token in the Authorization header:
        `Authorization: Bearer <token>`

  # ----------------------------------------------------------
  # Reusable Parameters
  # ----------------------------------------------------------
  parameters:
...
    startTime:
      name: start_time
      in: query
      description: Start of time range (RFC 3339)
      schema:
        type: string
        format: date-time
    endTime:
      name: end_time
      in: query
      description: End of time range (RFC 3339)
      schema:
        type: string
        format: date-time
    limit:
      name: limit
      in: query
      description: Results per page
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 1000
    offset:
      name: offset
      in: query
      description: Page offset (number of results to skip)
      schema:
        type: integer
        default: 0
        minimum: 0
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Forbidden — valid read token but write token required for this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  # ----------------------------------------------------------
  # Schemas
  # ----------------------------------------------------------
  schemas:
...
    CallState:
      type: string
      enum: [monitoring, recording, stopped, completed]
      description: |
        Call recording state (mapped from trunk-recorder integer values):
        - **monitoring**: control channel activity seen, no audio capture yet
        - **recording**: actively capturing audio
        - **stopped**: recording ended, finalizing
        - **completed**: audio file written, ready for access
    MonState:
      type: string
      enum: [unspecified, active, duplicate]
      description: |
        Call monitor state (from trunk-recorder):
        - **unspecified**: default / not specifically monitored
        - **active**: actively monitored
        - **duplicate**: duplicate call on another recorder
...
    Error:
      type: object
      required: [code, error]
      properties:
        code:
          type: string
          description: Machine-readable error code for programmatic handling
          example: not_found
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - rate_limited
            - internal_error
            - service_unavailable
            - invalid_body
            - invalid_parameter
            - invalid_time_range
            - query_failed
            - ambiguous_id
            - duplicate
            - request_timeout
        error:
          type: string
          example: Resource not found
        detail:
          type: string
          description: Additional context when available
...
    Call:
      type: object
      description: A recorded radio call/transmission
      required: [call_id, system_id, tgid, start_time]
      properties:
        # Identifiers
        call_id:
          type: integer
          description: Unique database ID for this recording
          example: 48531
        call_group_id:
          type: integer
          description: |
            Call group this recording belongs to. Multiple recordings of the
            same transmission (from different sites) share a call_group_id.
            Derived from (system_id, tgid, start_unix).
          example: 1

        # System context (embedded to avoid lookup)
        system_id:
          type: integer
          description: Logical radio system this call belongs to
          example: 1
        system_name:
          type: string
          description: System display name
          example: "Butler/Warren P25"
        sysid:
          type: string
          description: P25 SYSID (for display; "0" for conventional)
          example: "348"

        # Site context (which recorder captured this)
        site_id:
          type: integer
          description: Recording site that captured this call
          example: 1
        site_short_name:
          type: string
          description: TR sys_name of the capturing site (e.g., "butco", "warco")
          example: butco

        # Talkgroup context (embedded to avoid lookup)
        tgid:
          type: integer
          example: 9178
        tg_alpha_tag:
          type: string
          description: Talkgroup short-form display name
          example: "09-8L Main"
        tg_description:
          type: string
          description: Human-readable talkgroup description
          example: Middletown Police Dispatch
        tg_tag:
          type: string
          description: Functional category tag (e.g., Law Dispatch, Fire Tac)
          example: Law Dispatch
        tg_group:
          type: string
          description: Broad organizational grouping
          example: "Butler County (09) Law"

        # Timing
        start_time:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        stop_time:
          type: string
          format: date-time
          nullable: true
          description: "Null for active (in-progress) calls"
          example: "2024-01-15T10:30:45Z"
        duration:
          type: number
          description: Duration in seconds (elapsed so far for active calls)
          example: 45.5

        # Audio
        audio_url:
          type: string
          nullable: true
          description: "Relative URL to audio endpoint. Null for active calls."
          example: /api/v1/calls/48531/audio
        audio_type:
          type: string
          description: "Audio encoding type from trunk-recorder (e.g., `digital`, `digital tdma`, `analog`)"
          example: digital
        audio_size:
          type: integer
          description: Audio file size in bytes
          example: 45000

        # Signal quality
        freq:
          type: integer
          description: Frequency in Hz
          example: 851012500
        freq_error:
          type: integer
          description: Frequency error in Hz
          example: 0
        signal_db:
          type: number
          nullable: true
          description: Signal strength in dB (null if unavailable; MQTT sends 999 as sentinel)
          example: -23
        noise_db:
          type: number
          nullable: true
          description: Noise floor in dB (null if unavailable; MQTT sends 999 as sentinel)
          example: -50
        error_count:
          type: integer
          description: Decode error count
          example: 2
        spike_count:
          type: integer
          description: Audio spike count
          example: 0

        # Status
        call_state:
          $ref: "#/components/schemas/CallState"
        mon_state:
          $ref: "#/components/schemas/MonState"
        emergency:
          type: boolean
          example: false
        encrypted:
          type: boolean
          example: false
        analog:
          type: boolean
          example: false
        conventional:
          type: boolean
          example: false
        phase2_tdma:
          type: boolean
          example: false
        tdma_slot:
          type: integer
          description: "TDMA slot (0=slot 1 or N/A, 1=slot 2)"
          example: 0

        # Patches
        patched_tgids:
          type: array
          description: Other talkgroup IDs patched into this call
          items:
            type: integer

        # Transmission and frequency data (JSONB columns)
        src_list:
          type: array
          description: |
            Source/transmission list from trunk-recorder's srcList.
            Each entry is a unit transmission segment within the call.
            Also available via GET /calls/{id}/transmissions.
          items:
            $ref: "#/components/schemas/CallTransmission"
        freq_list:
          type: array
          description: |
            Frequency list from trunk-recorder's freqList.
            Each entry is a frequency segment within the call.
            Also available via GET /calls/{id}/frequencies.
          items:
            $ref: "#/components/schemas/CallFrequency"
        unit_ids:
          type: array
          description: Denormalized array of distinct unit radio IDs (src values) from src_list. GIN-indexed for fast unit filtering.
          items:
            type: integer

        # Units on this call (embedded)
        units:
          type: array
          items:
            $ref: "#/components/schemas/CallUnit"

        # Transcription (denormalized from transcriptions table)
        has_transcription:
          type: boolean
          description: Whether this call has been transcribed
          example: true
        transcription_status:
          $ref: "#/components/schemas/TranscriptionStatus"
        transcription_text:
          type: string
          nullable: true
          description: Primary transcription text (null if no transcription)
          example: Engine 5 responding to the scene
        transcription_word_count:
          type: integer
          nullable: true
          example: 6

        # Extensible metadata
        metadata_json:
          type: object
          additionalProperties: true
          description: Arbitrary metadata from trunk-recorder
    CallUnit:
      type: object
      description: A unit that transmitted during a call
      required: [system_id, unit_id]
      properties:
        system_id:
          type: integer
          description: Database system ID for disambiguation in multi-system deployments
          example: 1
        unit_id:
          type: integer
          description: Unit radio ID (RID)
          example: 924003
        alpha_tag:
          type: string
          description: Unit display name
          example: "09 7COM3"
...
    CallFrequency:
      type: object
      description: |
        A frequency segment from trunk-recorder's freqList. Each entry
        represents a period of time the call spent on a given frequency.
      properties:
        freq:
          type: integer
          description: Frequency in Hz
          example: 154875000
        time:
          type: string
          format: date-time
          description: When this frequency segment started (RFC 3339)
          example: "2024-04-15T19:03:22Z"
        pos:
          type: number
          description: Position in the audio file in seconds
          example: 0.0
        len:
          type: number
          description: Duration on this frequency in seconds
          example: 3.24
        error_count:
          type: integer
          example: 50
        spike_count:
          type: integer
          example: 3
    CallTransmission:
      type: object
      description: |
        A unit key-up from trunk-recorder's srcList. Each entry represents
        one unit's transmission within the call. The `src` field maps to
        the unit radio ID (RID).
      properties:
        src:
          type: integer
          description: Unit radio ID (RID) — matches unit_id in other models
          example: 104
        tag:
          type: string
          description: Unit alpha tag at time of transmission (from trunk-recorder)
          example: "09 7COM3"
        time:
          type: string
          format: date-time
          description: When this unit keyed up (RFC 3339)
          example: "2024-04-15T19:03:22Z"
        pos:
          type: number
          description: Position in audio file in seconds
          example: 0.0
        duration:
          type: number
          description: Duration of this transmission segment in seconds (computed from gap to next transmission or end of call)
          example: 3.5
        emergency:
          type: integer
          description: "Emergency flag (0=normal, 1=emergency)"
          example: 0
        signal_system:
          type: string
          description: Signal system identifier (typically empty)
          example: ""
    CallGroup:
      type: object
      description: |
        A group of duplicate call recordings from multiple recorders.
        Derived from (system_id, tgid, start_unix) — recordings that match
        on all three are grouped.
      properties:
        id:
          type: integer
          example: 1
        # System context (embedded)
        system_id:
          type: integer
          example: 1
        system_name:
          type: string
          description: System display name
          example: "Butler/Warren P25"
        sysid:
          type: string
          description: P25 SYSID (for display)
          example: "348"
        # Site that captured the primary recording
        site_id:
          type: integer
          example: 1
        site_short_name:
          type: string
          example: butco
        # Talkgroup context (embedded)
        tgid:
          type: integer
          example: 9178
        tg_alpha_tag:
          type: string
          description: Talkgroup short-form display name
          example: "09-8L Main"
        tg_description:
          type: string
          description: Human-readable talkgroup description
          example: Middletown Police Dispatch
        tg_tag:
          type: string
          description: Functional category tag
          example: Law Dispatch
        tg_group:
          type: string
          description: Broad organizational grouping
          example: "Butler County (09) Law"
        # Timing
        start_time:
          type: string
          format: date-time
        stop_time:
          type: string
          format: date-time
        duration:
          type: number
          example: 45.5
        # Group info
        call_count:
          type: integer
          description: Number of recordings in this group
          example: 2
        primary_call_id:
          type: integer
          description: Call ID of the best-quality recording
          example: 48531
        # Transcription (denormalized from primary call)
        has_transcription:
          type: boolean
          example: true
        transcription_status:
          $ref: "#/components/schemas/TranscriptionStatus"
        transcription_text:
          type: string
          nullable: true
          example: Engine 5 responding to the scene

    # ========== Transcription schemas ==========
    TranscriptionStatus:
      type: string
      enum: [none, auto, reviewed, verified, excluded]
      description: |
        - **none**: no transcription attempted
        - **auto**: machine-transcribed, unreviewed
        - **reviewed**: passed quality checks
        - **verified**: human-verified correct
        - **excluded**: permanently excluded from dataset
...
    SystemMergeRequest:
      type: object
      required: [source_id, target_id]
      properties:
        source_id:
          type: integer
          description: System to merge FROM (will be deleted after merge)
          example: 3
        target_id:
          type: integer
          description: System to merge INTO (keeps its system_id)
          example: 1
        update_target_metadata:
          type: boolean
          description: |
            If true, copy short_name/sysid/wacn/nac from source to target
            (useful when the source has better metadata, e.g., was identified
            after the target was created as "unknown").
          default: false
    SystemMergeResponse:
      type: object
      properties:
        target_id:
          type: integer
          description: The surviving system_id
          example: 1
        source_id:
          type: integer
          description: The deleted system_id
          example: 3
        calls_moved:
          type: integer
          description: Number of call records repointed
          example: 5200
        talkgroups_moved:
          type: integer
          description: Talkgroups moved from source (source-only)
          example: 12
        talkgroups_merged:
          type: integer
          description: Talkgroups that existed in both (stats combined)
          example: 45
        units_moved:
          type: integer
          example: 30
        units_merged:
          type: integer
          example: 150
        events_moved:
          type: integer
          description: Unit events repointed
          example: 25000
...
    CallGroupListResponse:
      type: object
      required: [call_groups, total, limit, offset]
      properties:
        call_groups:
          type: array
          items:
            $ref: "#/components/schemas/CallGroup"
        total:
          type: integer
          example: 500
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 0
    CallGroupDetailResponse:
      type: object
      required: [call_group, calls]
      properties:
        call_group:
          $ref: "#/components/schemas/CallGroup"
        calls:
          type: array
          description: All individual recordings in this group
          items:
            $ref: "#/components/schemas/Call"
...